2022-2-VK-QA-PYTHON-dimasikanz-main/requirements.txt
allure-pytest==2.11.0
faker==15.0.0
pytest==7.1.3
pytest-xdist==2.5.0
selenium==4.5.0
webdriver_manager==3.8.3
requests==2.28.1